FIXED #169: Merge frame processor functions
